dies on insert into database see TODO.

+++ app/src/main/java/fuelfinder/mann/Utility/MileageModelDataSource.java
@@ -30,6 +30,7 @@ public class MileageModelDataSource {
         dbHelper = new MySQLiteHelper(context);
     }
 
+
     public void open() throws SQLException {
         database = dbHelper.getWritableDatabase();
     }
@@ -38,13 +39,15 @@ public class MileageModelDataSource {
         dbHelper.close();
     }
 
-    public MileageModel createMileageModel(int Engine, String Make, int Mileage, String Model, String Name, int VehicleID, String VehicleName, int Year, String Transmission) {
+    public MileageModel createMileageModel(double Engine, String Make, double Mileage, String Model, String Name, int VehicleID, String VehicleName, int Year, String Transmission) {
         ContentValues values = new ContentValues();
         values.put(MySQLiteHelper.VehicleID, VehicleID);
         long insertId = database.insert(MySQLiteHelper.NAME, null,
                 values);
         ContentValues engValues = new ContentValues();
         engValues.put(MySQLiteHelper.Engine, Engine);
+      //TODO:  unknown death at this point
+
         long insertEngine = database.insert(MySQLiteHelper.NAME, null, engValues );
 
         ContentValues makeValues = new ContentValues();
